+++ tests/test_decorator_input.py
@@ -11,7 +11,10 @@ from .schemas import Files
 from .schemas import Foo
 from .schemas import Form
 from .schemas import FormAndFiles
+from apiflask import Schema
 from apiflask.fields import String
+from apiflask.validators import Length
+from apiflask.validators import OneOf
 
 
 def test_input(app, client):
@@ -413,3 +416,136 @@ def test_input_body_example(app, client):
         rv.json['paths']['/baz']['post']['requestBody']['content']['application/json']['examples']
         == examples
     )
+
+
+def test_skip_validation(app, client):
+    incorrect_json = {'name': 'Kitty', 'category': 'unknown'}
+
+    class PetIn(Schema):
+        name = String(required=True, validate=Length(0, 10))
+        category = String(required=True, validate=OneOf(['dog', 'cat']))
+
+    @app.patch('/pets_without_validation/<int:pet_id>')
+    @app.input(PetIn, validation=False)
+    def pets_without_validation(pet_id, json_data):
+        return {'pet_id': pet_id, 'json_data': json_data}
+
+    no_validated_rv = client.patch('/pets_without_validation/1', json=incorrect_json)
+    assert no_validated_rv.status_code == 200
+    assert no_validated_rv.json['json_data']['name'] == 'Kitty'
+    assert no_validated_rv.json['json_data']['category'] == 'unknown'
+
+
+def test_location_raw_data(app):
+    from apiflask.scaffold import _load_raw_data
+
+    # will raises ValueError
+    location = 'unknown'
+    with pytest.raises(ValueError, match=f'Invalid location argument: {location}'):
+        _load_raw_data('unknown')
+
+    # `json` location
+    with app.test_request_context('/load_raw_data', method='POST', json={'name': 'Kitty'}):
+        json_data = _load_raw_data()
+        assert json_data['name'] == 'Kitty'
+
+    with app.test_request_context('/load_raw_data', method='POST', data={'name': 'Kitty'}):
+        json_data = _load_raw_data()
+        assert json_data == {}
+
+    # `form` location
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+        data={'username': 'Kitty', 'password': '<PASSWORD>'},
+        content_type='application/x-www-form-urlencoded',
+    ):
+        form_data = _load_raw_data('form')
+        assert form_data['username'] == 'Kitty'
+        assert form_data['password'] == '<PASSWORD>'
+
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+        json={'username': 'Kitty', 'password': '<PASSWORD>'},
+    ):
+        form_data = _load_raw_data('form')
+        assert form_data == {}
+
+    # `files` or `form_and_files` location
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+        data={'file': (io.BytesIO(b'Hello World!'), 'demo.txt')},
+        content_type='multipart/form-data',
+    ):
+        files_data = _load_raw_data('files')
+        assert files_data['file'].filename == 'demo.txt'
+        assert files_data['file'].stream.read() == b'Hello World!'
+
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+        data={'name': 'Kitty', 'file': (io.BytesIO(b'Hello World2!'), 'demo2.txt')},
+        content_type='multipart/form-data',
+    ):
+        files_and_form_rv = _load_raw_data('form_and_files')
+        assert files_and_form_rv['file'].filename == 'demo2.txt'
+        assert files_and_form_rv['file'].stream.read() == b'Hello World2!'
+        assert files_and_form_rv['name'] == 'Kitty'
+
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+        json={'username': 'Kitty', 'password': '<PASSWORD>'},
+    ):
+        form_and_files_data = _load_raw_data('form_and_files')
+        assert form_and_files_data == {}
+
+    # `query` location
+    with app.test_request_context(
+        '/load_raw_data?name=Kitty',
+        method='GET',
+    ):
+        query_data = _load_raw_data('query')
+        assert query_data == {'name': 'Kitty'}
+
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+    ):
+        query_data = _load_raw_data('query')
+        assert query_data == {}
+
+    # `json_or_form`
+    with app.test_request_context(
+        '/load_raw_data',
+        data={'username': 'Kitty', 'password': '<PASSWORD>'},
+        content_type='application/x-www-form-urlencoded',
+    ):
+        json_or_form_rv = _load_raw_data('json_or_form')
+        assert json_or_form_rv['username'] == 'Kitty'
+        assert json_or_form_rv['password'] == '<PASSWORD>'
+
+    with app.test_request_context(
+        '/load_raw_data',
+        json={'currentId': '<current_id>'},
+    ):
+        json_or_form_data = _load_raw_data('json_or_form')
+        json_or_form_data['currentId'] = '<current_id>'
+
+    with app.test_request_context(
+        '/load_raw_data',
+        method='POST',
+    ):
+        json_or_form_data = _load_raw_data('json_or_form')
+        assert json_or_form_data == {}
+
+    # `path` location
+    @app.get('/load_raw_data/<int:pet_id>')
+    def load_raw_data(pet_id):
+        return _load_raw_data('path')
+
+    with app.test_request_context('/load_raw_data/1', method='GET'):
+        path_data = app.dispatch_request()
+        assert path_data == {'pet_id': 1}
